LS-007-A implement new Test, add Login Layout [LB,RT]
