+++ app/src/main/java/com/example/loginsesame/helper/LogTag.kt
@@ -1,5 +1,6 @@
 package com.example.loginsesame.helper
 
 class LogTag {
-    val LOG_MAIN = "LogActivity"
+    val LOG_MAIN = "MainActivity"
+    val LOG_LOGIN = "LoginActivity"
 }
