+++ local.properties
@@ -1,8 +1,10 @@
-## This file must *NOT* be checked into Version Control Systems,
+## This file is automatically generated by Android Studio.
+# Do not modify this file -- YOUR CHANGES WILL BE ERASED!
+#
+# This file should *NOT* be checked into Version Control Systems,
 # as it contains information specific to your local configuration.
 #
 # Location of the SDK. This is only used by Gradle.
 # For customization when using a Version Control System, please read the
 # header note.
-#Sun Apr 18 11:22:49 CEST 2021
-sdk.dir=/home/bernhard/Android/Sdk
+sdk.dir=/Users/treith/Library/Android/sdk
